requirements.txt
beautifulsoup4
Flask
Flask-SQLAlchemy
psycopg2
python-dotenv
requests
SQLAlchemy